mkisofs/mkisofs.c: Joliet is no longer nonstandard
With the 2019 revision of ECMA 119, Joliet has been added as an annex to
the standard.  Hence, the warning about Joliet being nonstandard is no
longer appropriate.  Neverthless, the warning is kept as the
recommendation to use Rockridge in conjunction with Joliet is still
sound.

Reported by:	Sam Trenholme <maradns@gmail.com>
Closes		#7