Add findlinks.sh, .gitignore and (re)generate .links
This script scans the various MKLINKS files and generates appropriate
.gitignore files in each affected directory.  At the same time, these
links are recorded in the .links file.

While we are at it, add the .gitignore files generated thus and refresh
the .links file.  Apart from getting sorted, this removes the stale
entry for ./sccs/sccs/cmd/xpg4/delta.c and adds a new entry for
./star/star_fat.c to the .links file.

A hand-written .gitignore file in the root directory is also added.
This file holds ignored files that are not covered by findlinks.sh as
well as generic patterns for build artifacts.