DEFAULTS/Defaults.freebsd: switch FreeBSD default link mode to dynamic
Dynamic link mode has been tested in the ports tree for over a year
and has worked without flaw so far.  Enable it as the default.

While we are at it, set INS_RBASE to be relative to INS_BASE as is
convention on FreeBSD; user applications are really not supposed to
put their configuration files into /etc.

We also change the default compiler to clang as it has been the
default on FreeBSD for a while.  Pcc has never been used on FreeBSD;
not sure why it was the default (it did work by coincidence though).