autoconf/aclocal.m4: work around clang bug in #pragma weak detection
clang really doesn't like #pragma weak on functions without prototypes.
So give a prototype to the symbol and make clang happy.  This does
unfortunately introduce a dependency on prototypes to the test, but
I suppose compilers without prototypes but with #pragma weak are rather
rare.

As a secondary workaround, move #pragma weak after all uses of the
functions in question as clang gets confused when you call a function
that has a weak alias defined with #pragma weak.