star: document {,de}compression helpers as unavailable on DOS
There's a separate code path for DJGPP which I cannot test right now due
to missing DJGPP packages on Debian and FreeBSD.  It would be great to
have this supported in a future version.

Also ensure that star_get_cmd_flags() and friends are not compiled in
when we don't actually ship the feature.