sh: Fix bug that causes shell to exit when -b / notify option was set
The list flagval list translates from an index to the bit that should
be set in the flags set. There are two sets of flags: flags and
flags2. To indicte in which of the sets the flag should be set, the
fl2 macro provided an indictor bit to tell. The notifyflg entry was
missing that bit and incorretly set the bit in flags instead of flags2
causing the shell to send itself a SIGINT upon typing ^C and thus exiting.