autoconf/os-sunos5.def: detect man style with configure test
Different make implementations have different macros to run shell
commands. Previously a shell command was used in such an unportable
macro call to detect the style of manual pages on a given Solaris
system. This broke when building with GNU make on such systems.

This patch moves the test to a configure script and instead substitutes
a make internal variable to indicate the Solaris manstyle.

Tested to work on OmniOS i86pc (Illumos/SunOS 5.11).

Fixes		#51
Reviewed by:	Robert Clausecker <fuz@fuz.su>
Signed-off-by:	Nico Sonack <nsonack@herrhotzenplotz.de>