Fix GNU make builds and iconv linkage on FreeBSD
The autoconf tests are sort of broken because of the inclusion order
of headers on FreeBSD. The GNU make issue was easy to solve and needed
some magic in a make include file. The include order was hacked to be
correct by ensuring that we first include the system headers and then
ports headers. This must be done in order to ensure we do not drag in
the headers for GNU libiconv in case it is installed because it does
some nasty ifdef magic and breaks linking against the FreeBSD internal
libc iconv routines.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>