sunpro/Make: remove register from C++ sources
The register keyword has been removed from the C++ language with C++17.
Discard the keyword from all C++ sources to enable a build with the new
standard.

This commit is a bulk deletion of the keyword without any fixes to
formatting and indentation.  A second commit will deal with that.

Fixes:		#13
Reviewed-by:	Nico Sonack <nsonack@herrhotzenplotz.de>