mkisofs/write.c: produce less scrollback when logging progress to tty
The mkisofs program produces quite a bit of progress output.
The output once had a somewhat reasonable pace, but on modern
computers, it produces a new line of text multiple times a
second.  Change things so that when GUI mode is disabled and
output goes to a TTY, we overwrite the progress line instead
of producing a new line if possible.

Fixes		#65
Reported-by	Michael Dexter <editor@callfortesting.org>
Reviewed-by	Nico Sonack <nsonack@herrhotzenplotz.de>